# [breaking] [vm] Remove hash in access path