--- a/language/diem-vm/src/lib.rs
+++ b/language/diem-vm/src/lib.rs
@@ -162,5 +162,5 @@ pub trait VMExecutor: Send {
 /// Get the AccessPath to a resource stored under `address` with type name `tag`
 fn create_access_path(address: AccountAddress, tag: StructTag) -> AccessPath {
     let resource_tag = ResourceKey::new(address, tag);
-    AccessPath::resource_access_path(&resource_tag)
+    AccessPath::resource_access_path(resource_tag)
 }