--- a/types/src/access_path.rs
+++ b/types/src/access_path.rs
@@ -51,38 +51,39 @@ pub struct AccessPath {
     pub path: Vec<u8>,
 }
 
-impl AccessPath {
-    pub const CODE_TAG: u8 = 0;
-    pub const RESOURCE_TAG: u8 = 1;
+#[derive(Clone, Eq, PartialEq, Hash, Serialize, Deserialize, Ord, PartialOrd)]
+pub enum Path {
+    Code(ModuleId),
+    Resource(StructTag),
+}
 
+impl AccessPath {
     pub fn new(address: AccountAddress, path: Vec<u8>) -> Self {
         AccessPath { address, path }
     }
 
-    pub fn resource_access_vec(tag: &StructTag) -> Vec<u8> {
-        tag.access_vector()
+    pub fn resource_access_vec(tag: StructTag) -> Vec<u8> {
+        lcs::to_bytes(&Path::Resource(tag)).expect("Unexpected serialization error")
     }
 
     /// Convert Accesses into a byte offset which would be used by the storage layer to resolve
     /// where fields are stored.
-    pub fn resource_access_path(key: &ResourceKey) -> AccessPath {
-        let path = AccessPath::resource_access_vec(&key.type_());
+    pub fn resource_access_path(key: ResourceKey) -> AccessPath {
+        let path = AccessPath::resource_access_vec(key.type_);
         AccessPath {
-            address: key.address().to_owned(),
+            address: key.address,
             path,
         }
     }
 
-    fn code_access_path_vec(key: &ModuleId) -> Vec<u8> {
-        key.access_vector()
+    fn code_access_path_vec(key: ModuleId) -> Vec<u8> {
+        lcs::to_bytes(&Path::Code(key)).expect("Unexpected serialization error")
     }
 
-    pub fn code_access_path(key: &ModuleId) -> AccessPath {
+    pub fn code_access_path(key: ModuleId) -> AccessPath {
+        let address = *key.address();
         let path = AccessPath::code_access_path_vec(key);
-        AccessPath {
-            address: *key.address(),
-            path,
-        }
+        AccessPath { address, path }
     }
 }