--- a/types/src/account_config/resources/preburn_balance.rs
+++ b/types/src/account_config/resources/preburn_balance.rs
@@ -41,7 +41,7 @@ impl PreburnResource {
 
     // TODO: remove this once the MoveResource trait allows type arguments to `resource_path`.
     pub fn access_path_for(currency_typetag: TypeTag) -> Vec<u8> {
-        AccessPath::resource_access_vec(&PreburnResource::struct_tag_for_currency(currency_typetag))
+        AccessPath::resource_access_vec(PreburnResource::struct_tag_for_currency(currency_typetag))
     }
 }