… language/ir-testsuite/tests/testing_infra/show_all_optional_info.exp
… language/move-lang/functional-tests/tests/testing-infra/show_all_optional_info.exp
… language/stdlib/compiled/packed_types/packed_types.txt
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::account_limits::account_limits.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::admin_script::admin_script_rotate_key_multi_signer.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::admin_script::admin_script_rotate_key_single_signer_new_epoch.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::admin_script::admin_script_rotate_key_single_signer_no_epoch.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::create_account::create_account.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::data_store::borrow_after_move.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::data_store::change_after_move.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::data_store::move_from_across_blocks.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::emergency_admin_script::halt_network.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::emergency_admin_script::validator_batch_remove.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::mint::mint_to_existing_not_dd.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::mint::mint_to_new_account.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::mint::tiered_mint_designated_dealer.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::mint::tiered_update_exchange_rate.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::module_publishing::bad_module_address.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::module_publishing::duplicate_module.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::module_publishing::test_publishing_allow_modules.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::module_publishing::test_publishing_no_modules_non_allowlist_script_proper_sender.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::module_publishing::test_publishing_no_modules_proper_sender.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::on_chain_configs::drop_txn_after_reconfiguration.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::on_chain_configs::initial_diem_version.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::on_chain_configs::update_script_allow_list.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::on_chain_configs::updated_limit_allows_txn.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::peer_to_peer::few_peer_to_peer_with_event.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::peer_to_peer::single_peer_to_peer_with_event.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::peer_to_peer::single_peer_to_peer_with_padding.exp
… language/testing-infra/e2e-tests/goldens/language_e2e_testsuite::tests::peer_to_peer::zero_amount_peer_to_peer.exp