@jolestar 2020-12-11
Is there a security problem if someone maliciously constructs a very long key through a nested Generic type structure?