@runtian-zhou 2020-12-11
> Is there a security problem if someone maliciously constructs a very long key through a nested Generic type structure?

That is actually a ligit concern. Our repsonse is following:
1. You pay for the data. We will be charged more if you are planning to put a longer name in storage.
2. It is another spot that the system will have to deal with perf implications. For the VM this could mean string interning or other classic string perf solutions

On the other hand, we strongly suggest that the drawbacks over there worth it for the benefits of not having the hash as this will make our storage totally self-describable.